Fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23332 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDialog/LogDlgFilter.cpp
@@ -128,7 +128,7 @@ bool CLogDlgFilter::Match (char* text, size_t size) const
                 // lowercase won't work: we first have to convert
                 // the string to wide-char. Very very slow, but unavoidable
                 // if we want to make this work right.
-                CStringA as = CStringA(text, size);
+                CStringA as = CStringA(text, (int)size);
                 CString s = CUnicodeUtils::GetUnicode(as);
                 s.MakeLower();
                 as = CUnicodeUtils::GetUTF8(s);
